# docs/swagger.yaml
definitions:
...
  dtos.CategoryDetailFormatter:
    properties:
      category_name:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      slug:
        type: string
      user:
        $ref: '#/definitions/dtos.CategoryUserFormatter'
      user_id:
        type: integer
    type: object
  dtos.CategoryFormatter:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      slug:
        type: string
      user_id:
        type: integer
    type: object
  dtos.CategoryUserFormatter:
    properties:
      full_name:
        type: string
      role:
        type: string
    type: object
...
  dtos.CreatePaymentCategoryInput:
    properties:
      payment_category_name:
        type: string
    type: object
...
  dtos.PaymentCategoryFormatter:
    properties:
      id:
        type: integer
      payment_category_name:
        type: string
      slug:
        type: string
      user:
        $ref: '#/definitions/dtos.PaymentCategoryUserFormatter'
      user_id:
        type: integer
    type: object
  dtos.PaymentCategoryUserFormatter:
    properties:
      full_name:
        type: string
      role:
        type: string
      user_id:
        type: integer
    type: object
...
  utils.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  utils.Response:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/utils.Meta'
    type: object
info:
  contact:
    email: nadapdaponai21@gmail.com
    name: API Support
    url: https://onainadapdap1.github.io/
  description: This is service for my Online Store API assignment.
  termsOfService: http://swagger.io/terms/
  title: Online Store API
  version: "1.0"
paths:
...
  /api/v1/categories:
    get:
      description: Get all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dtos.CategoryDetailFormatter'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get all categories
      tags:
      - categories
  /api/v1/categories/category:
    post:
      consumes:
      - multipart/form-data
      description: Create a new Category with a given name, description and image
        file
      parameters:
      - description: Name of the category
        in: formData
        name: name
        required: true
        type: string
      - description: Description of the photo
        in: formData
        name: description
        required: true
        type: string
      - description: Image file of the photo
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CategoryFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create Category
      tags:
      - categories
  /api/v1/categories/category/{id}:
    delete:
      description: Delete category by id
      parameters:
      - description: delete category by id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete category by id
      tags:
      - categories
  /api/v1/categories/category/{slug}:
    get:
      description: Get one photo by slug
      parameters:
      - description: get photo by slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CategoryFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get one photo by slug
      tags:
      - categories
    put:
      consumes:
      - application/json
      - multipart/form-data
      description: Update category
      parameters:
      - description: update category by slug
        in: path
        name: slug
        required: true
        type: string
      - description: name of the category to be updated
        in: formData
        name: name
        required: true
        type: string
      - description: description of the category to be updated
        in: formData
        name: description
        required: true
        type: string
      - description: image file of the category to be updated
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.CategoryFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - categories
...
  /api/v1/paymentcategories:
    post:
      consumes:
      - application/json
      description: Create Payment Category
      parameters:
      - description: Payment category input
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dtos.CreatePaymentCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PaymentCategoryFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/utils.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create Payment Category
      tags:
      - payment-categories
  /api/v1/paymentcategories/paymentcategory/{slug}:
    get:
      description: Get payment category by slug
      parameters:
      - description: Get payment category by slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dtos.PaymentCategoryFormatter'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get payment category by slug
      tags:
      - payment-categories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"